requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch~=2.6.0
transformers~=4.48.3
clickhouse-connect==0.6.21
python-dotenv==1.0.0
pydantic==2.10.6
prometheus-client==0.19.0
fastapi[all]~=0.109.1
numpy~=1.26.4
pandas==2.1.4
tqdm~=4.67.1
hdbscan==0.8.40
scikit-learn~=1.6.0
sqlalchemy~=2.0.38
pytz~=2024.2
usearch==2.16.9
apscheduler==3.10.4
jinja2~=3.1.4
openai~=1.61.1
faker~=33.1.0
starlette~=0.36.3
uvicorn~=0.34.0
gunicorn==23.0.0
kafka-python~=2.0.3
pymysql==1.1.1
cryptography==44.0.0
langchain==0.3.19
langchain-core==0.3.37
langchain-openai==0.3.6